requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2==2.9.9
kenar==0.4.3
pydantic==2.7.4
httpx==0.27.0
